assignments/warm-up/maayan-babayof/requirements.txt
fastapi
uvicorn
pytest
httpx
mangum
requests
beautifulsoup4
python-dateutil
